requirements.txt
Django==4.2.9
django-jazzmin==3.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-environ==0.11.2
Pillow==10.1.0
django-filter==23.5
iyzipay==1.0.45
stripe==8.0.0
twilio==9.0.0
django-phonenumber-field==7.3.0
phonenumbers==8.13.27
django-environ==0.11.2
gunicorn==21.2.0
whitenoise==6.6.0
coverage==7.4.0
faker==22.0.0
freezegun==1.4.0
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
boto3==1.34.51
django-storages==1.14.2
